Try to quiet warnings, #3250
git-svn-id: http://svn.osgeo.org/postgis/trunk@13956 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/cunit/cu_algorithm.c
@@ -886,7 +886,7 @@ static void test_geohash_point_as_int(void)
 {
 	unsigned int gh;
 	POINT2D p;
-	unsigned int rs;
+	unsigned long long rs;
 	
 	p.x = 50; p.y = 35;
 	gh = geohash_point_as_int(&p);
